+++ CGoGN/include/Algo/Render/GL2/topo3Render.hpp
@@ -820,7 +820,7 @@ void Topo3RenderGMap<PFP>::updateData(MAP& mapx, const VertexAttribute<VEC3, MAP
 			dd = mapx.phi1(dd);
 		} while (dd != d);
 		
-		unsigned int nb = vecPos.size();
+		unsigned int nb = uint32(vecPos.size());
 		
 		vecPos.push_back(vecPos.front()); // copy the first for easy computation on next loop
